Initial version of mvm_quality_graphics.py module   (#1408)
* Initial version of mvm_quality_graphics.py module.  Infrastructure setup and
initial implementation of vertical bar chart depicting WCSNAMEs per layer of a skycell.
This is a Work In Progress - need proper mapping of colors to WCSNAMEs to be
hard set.

* Force all possible WCSNAMEs to map to a specific color so the WCSNAME will
always be presented in the same color.

* Fixed the plot overwrite and generate plots in a grid (2 columns and
multiple rows) as it is not known a priori how many datasets will be
processed.  It may be the display of multiple plots should be done
differently.