Added try/except blocks for exceptions generated with regions do not have cross matches (#1419)
* Added try/except block for a KeyError, generated for a dictionary, and
an IndexError, generated for a 2D mask, where information may be missing
as the regions do not overlap or do not overlap enough to have the data
to populate the data structures.  Tested with skycell-p1637x16y01 which
has 24 input SVM images over three visits, four filters, and two detectors.

* Added a warning message at the beginning of processing that the SVM catalogs
for the component visits need to be present for proper quality assessment of
the MVM product.