Added MVM WCS graphics to mvm_quality_graphics module (#1421)
* Removed FIT_EVM_* WCS names which currently live in both mvm_quality_analysis and
mvm_quality_grahics modules.  Added the generate_wcs_footprint function and helper
function compute_overlap_stats.

* Added the WCSName mismatch footprint graphic and both graphics for
the footprints use the Spectral11 colormap.

* Added a legend for each skycell output page.  Removed triple quotes for
code temporarily meant to be skipped.

* Share the wcs_pref_list between the mvm_quality_analysis and mvm_quality_graphics
modules. WIP

* Do not specify the portion of the image to display as it causes problems
with the actual image render.  The CNT image was not fully displayed though
the viewer could not tell.  The bug was found when comparing againt FITS
images generated for the same data.

* Scale the count overlay image by slicing.  We just want a smaller
image and do not want to modify the pixel intensities.