SVM: Fix logic for catalogs and cosmic ray checking (#1450)
* The datasets quoted in this ticket which indicate they are good for
testing a fix no longer even execute the apparent troublesome portion
of the code.  The entirety of the drizzlepac code has changed a great
deal in two years.  Since this has not been noted to have been an
issue and no longer fails for the datasets suggested, I will make a
modest change in the code which covers the case of zero sources. Note
that zero sources should not be able to reach this portion of the code
any longer.

* Switched order of conditions in "if" statement based on code review.