Cleanup SCI header (FLT/C, DRZ/C) keywords remaining from an 'a aposteriori' active WCS solution (#1465)
* Check the SCI extension(s) of the output FLT/FLC and DRZ/DRC files.  If the active
WCS solution is 'a priori', there should not be any valid values for NMATCHES,
RMS_RA/RMS_DEC, FITGEOM, and CRDER1/CRDER2 as these keywords are associated with
'a posteriori' solutions.  Ensure the WCSTYPE is based upon the active WCSNAME.
Clarify some comments.

* Added description of latest change to the CHANGELOG.

* Check the SCI extension(s) of the output FLT/FLC and DRZ/DRC files.  If the active
WCS solution is 'a priori', delete the following keywords if they are associated
with the active WCS as they are residue from a previous 'a posteriori' solution:
NMATCHES, RMS_RA/RMS_DEC, FITGEOM, and CRDER1/CRDER2. Ensure the WCSTYPE is based
upon the active WCSNAME to clean up any confusion.

* Update drizzlepac/hapsequencer.py

Co-authored-by: Mihai Cara <mcara@users.noreply.github.com>

Co-authored-by: Mihai Cara <mcara@users.noreply.github.com>