WFPC2: Issue a specific error code and create an empty manifest file for bad data (#1550)
* If analysis of a WFPC2 exposure determines the data cannot be processed
for some reason, the converted WFPC2 FLT file is currently deleted.  The
process will now exist with the return code of NO_VIABLE_DATA (65), and
an empty manifest file will be created. Also, fixed the option string,
added a little documentation, and added a BaseException to handle the
sys.exit().

* Updated the changelog.