Implemented an additional and more robust RMS determination for the background based upon the Median Absolute Deviation (MAD) algorithm. The MAD algorithm is now one of three ways the RMS is computed under the umbrella of computing the background of the input image: standard RMS generated by sigma_clipped_stats, RMS based upon image characteristics (read noise, no. of exposures, and exposure time), and the MAD algorithm.  The largest of the RMS values is ultimately used for further computation. Added or improved comments.
Removed the "bthresh" variable in the "dao" section of all the
instrument/detector "catalog" JSON files as an obsolete
entry.  Updated the "bthresh" variable in the "dao" section of
the ACS WFC "quality" JSON file to be "5.0" which is the same
as the other detectors and is the minimum for a stronger source.