Resolved the issue of duplicate "ID"s in the rows of the Total Point catalog. For "point" source identification, looping is done over a list of weight masks, tp_masks, when invoking the "finder" algorithms. These algorithms return tables with a unique "id" number for each row/source in the table. When tp_masks > 1, the returned tables are "vstacked" to generate a final table. Unfortunately, the "id" number was not updated to reflect the stacking of the tables which created rows with the same "id".  This error did not cause the code to fail, but it did generate a garbled table.
It is believed this bug was introduced in HLA-1407/PR#1939 which is an update for this
latest build (v3.10x) when the computation of the tp_masks was fixed as previously only
one mask was ever generated (or so it is believed).  In essence, this was a bug waiting
to happen.