Added an internal comment and moved two deprecated modules, make_cfgobj_files.py and run_config_utils.py, to the devutils subdirectory.
Updated the catalog configuration files:
1) Removed the deprecated variable, fwhm, from the sourcex object
2) Removed the deprecated variable, TWEAK_THRESHOLD, from the dao object
3) Moved the variable (key/value pair), "TWEAK_FWHMPSF": <float>, to the
   general object portion of the file as it applies to both catalogs

Modified the catalog_utils.py module to read the TWEAK_FWHMPSF variable
from the proper location in the catalog JSON file.  The Segmentation
catalog now reports the proper value for the Gaussian Filter FWHM which
is used to smooth the total detection image to enhance peak or
multi-scale detection.